refactor to make test easier
